requirements.txt
Django==4.2.3
python-dotenv==1.0.0
dj-database-url==2.1.0
Pillow==10.0.0
gunicorn==21.2.0
django-extensions==3.2.3
psycopg2==2.9.7
psycopg2-binary==2.9.7
django-jazzmin==2.6.0
fontawesomefree==6.4.2
django-ckeditor==6.7.0
django-countries==7.6.1
markdown==3.4.4
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
django-cors-headers==4.2.0
drf_yasg==1.21.7
email-verifier==1.0.0
geoip2==4.7.0
django-ipware==5.0.0
phonenumbers==8.13.23
django-phonenumber-field==7.2.0
django-colorfield==0.11.0
user-agents==2.2.0
django-tinymce==4.1.0
Faker==37.1.0
django-braces==1.17.0